requirements.txt
matrix-nio[e2e]
aiohttp[speedups]
cinemagoer
lxml
pillow
pilmoji
